Moved BlockItemStore tests to the appropriate file; minor style and typing improvements
